refactor(gglib-runtime): extract startup_guard tests to separate file
